# PeaRL_spec/openapi/pearl-api-v1.1.yaml
openapi: 3.1.0
info:
  title: PeaRL API
  version: 1.1.0
  description: API-first risk orchestration platform for autonomous coding and secure/responsible
    delivery.
servers:
- url: http://localhost:8080
  description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: ^proj_[A-Za-z0-9_-]+$
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 200
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    CompiledContextPackage:
      $ref: ../schemas/context/compiled-context-package.schema.json
...
    FindingsIngestRequest:
      $ref: ../schemas/findings/findings-ingest-request.schema.json
...
    ApprovalRequest:
      $ref: ../schemas/workflow/approval-request.schema.json
    ApprovalDecision:
      $ref: ../schemas/workflow/approval-decision.schema.json
...
    JobStatus:
      $ref: ../schemas/workflow/job-status.schema.json
...
    ErrorResponse:
      $ref: ../schemas/common/error-response.schema.json
  responses:
    BadRequest:
      description: Validation or contract error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
- bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}/compile-context:
    post:
      tags:
      - ContextCompile
      summary: Compile layered context into project artifacts and canonical package
      operationId: compileContext
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                schema_version:
                  type: string
                project_id:
                  type: string
                compile_options:
                  type: object
                  additionalProperties: false
                  properties:
                    write_repo_artifacts:
                      type: boolean
                    output_repo_path:
                      type: string
                    include_human_readable_summary:
                      type: boolean
                    strict_conflict_mode:
                      type: boolean
                    apply_active_exceptions:
                      type: boolean
                artifact_targets:
                  type: array
                  items:
                    type: string
                trace_id:
                  type: string
              required:
              - schema_version
              - project_id
              - compile_options
              - trace_id
            examples:
              compile:
                externalValue: ../examples/compile/compile-context.request.json
      responses:
        '202':
          description: Compile job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
              examples:
                accepted:
                  externalValue: ../examples/compile/compile-context.accepted.response.json
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/projects/{project_id}/compiled-package:
    get:
      tags:
      - ContextCompile
      summary: Get latest compiled context package
      operationId: getCompiledPackage
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TraceId'
      responses:
        '200':
          description: Compiled package
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompiledContextPackage'
              examples:
                compiled:
                  externalValue: ../examples/compile/compiled-package.response.json
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/findings/ingest:
    post:
      tags:
      - Findings
      summary: Ingest findings from external tools
      operationId: ingestFindings
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindingsIngestRequest'
            examples:
              ingest:
                externalValue: ../examples/findings/findings-ingest.request.json
      responses:
        '202':
          description: Batch accepted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  schema_version:
                    type: string
                  batch_id:
                    type: string
                  accepted_count:
                    type: integer
                  quarantined_count:
                    type: integer
                  normalized_count:
                    type: integer
                  job_id:
                    type: string
                  trace_id:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                required:
                - schema_version
                - batch_id
                - accepted_count
                - quarantined_count
                - trace_id
                - timestamp
              examples:
                accepted:
                  externalValue: ../examples/findings/findings-ingest.response.json
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/approvals/requests:
    post:
      tags:
      - Approvals
      summary: Create approval request
      operationId: createApprovalRequest
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalRequest'
            examples:
              approvalReq:
                externalValue: ../examples/approvals/create-approval.request.json
      responses:
        '201':
          description: Approval request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalRequest'
  /api/v1/approvals/{approval_request_id}/decide:
    post:
      tags:
      - Approvals
      summary: Record approval decision
      operationId: decideApproval
      parameters:
      - name: approval_request_id
        in: path
        required: true
        schema:
          type: string
          pattern: ^appr_[A-Za-z0-9_-]+$
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalDecision'
            examples:
              decision:
                externalValue: ../examples/approvals/decision.request.json
      responses:
        '200':
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDecision'
              examples:
                decision:
                  externalValue: ../examples/approvals/decision.response.json
...